requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
Django==4.1.6
djangorestframework==3.14.0
drf-spectacular==0.25.1
drf-standardized-errors==0.12.5
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.0
pluggy==1.0.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
